rest api can now filter out by zip code
