+++ pom.xml
@@ -116,6 +116,11 @@
             <artifactId>json</artifactId>
             <version>20140107</version>
         </dependency>
+
+        <!--<dependency>-->
+            <!--<groupId>org.springframework.boot</groupId>-->
+            <!--<artifactId>spring-boot-starter-data-rest</artifactId>-->
+        <!--</dependency>-->
     </dependencies>
