+++ src/main/java/com/bexarair/demo/repositories/AQRestRepository.java
@@ -5,10 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Date;
+import java.util.List;
 
 @Repository
-public interface ARRecordRepository extends JpaRepository <AirQualityRecord, Long>  {
-    AirQualityRecord findByZipCode(String zipCode);
+public interface AQRestRepository extends JpaRepository <AirQualityRecord, Long>  {
+    List<AirQualityRecord> findByZipCode(String zipCode);
     AirQualityRecord findByHourObserved(String hour);
     AirQualityRecord findByZipCodeAndHourObserved(String zipCode, String HourObserved);
     AirQualityRecord findByZipCodeAndDateObserved(String zipCode, Date dateObserved);
